Removed types/methods associated with older http.Response implementation
