+++ fs/route.go
@@ -1,7 +1,6 @@
 package fs
 
 import (
-	"io"
 	"io/ioutil"
 	"os"
 	"path"
@@ -10,23 +9,23 @@ import (
 )
 
 func NewRoute(contentRootPath string) http.Route {
-	return &NewFileSystemRoute{
+	return &FileSystemRoute{
 		ContentRootPath: contentRootPath,
 		Factory:         &LocalResources{},
 	}
 }
 
-type NewFileSystemRoute struct {
+type FileSystemRoute struct {
 	ContentRootPath string
 	Factory         ResourceFactory
 }
 
-func (route NewFileSystemRoute) Route(requested http.RequestMessage) http.Request {
+func (route FileSystemRoute) Route(requested http.RequestMessage) http.Request {
 	resource := route.resolveResource(requested)
 	return requested.MakeResourceRequest(resource)
 }
 
-func (route NewFileSystemRoute) resolveResource(requested http.RequestMessage) http.Resource {
+func (route FileSystemRoute) resolveResource(requested http.RequestMessage) http.Resource {
 	resolvedPath := path.Join(route.ContentRootPath, requested.Path())
 	info, err := os.Stat(resolvedPath)
 	if err != nil {
@@ -39,24 +38,17 @@ func (route NewFileSystemRoute) resolveResource(requested http.RequestMessage) h
 	}
 }
 
+func readFileNames(files []os.FileInfo) []string {
+	fileNames := make([]string, len(files))
+	for i, file := range files {
+		fileNames[i] = file.Name()
+	}
+
+	return fileNames
+}
+
 type ResourceFactory interface {
 	DirectoryListingResource(message http.RequestMessage, files []string) http.Resource
 	ExistingFileResource(message http.RequestMessage, path string) http.Resource
 	NonExistingResource(message http.RequestMessage) http.Resource
-}
-
-type FileSystemRoute struct {
-	ContentRootPath string
-	Resource        FileSystemResource
-}
-
-func (route FileSystemRoute) Route(requested http.RequestMessage) http.Request {
-	return requested.MakeResourceRequest(route.Resource)
-}
-
-// Represents files and directories on the file system
-type FileSystemResource interface {
-	Name() string
-	Get(client io.Writer, message http.RequestMessage)
-	Head(client io.Writer, message http.RequestMessage)
 }
