+++ main/cmd/factory_test.go
@@ -69,10 +69,6 @@ var _ = Describe("InterruptFactory", func() {
 		})
 	})
 
-	Describe("ResourceFactory methods", func() {
-		XIt("implements ResourceFactory")
-	})
-
 	Describe("TCPServer", func() {
 		var (
 			server      cmd.Server
